adds v2 of models
